+++ src/services/webappclientservice.cpp
@@ -139,7 +139,7 @@ void WebAppClientService::initClipboardService() {
     QClipboard *clipboard = QApplication::clipboard();
 
     // React to clipboard changes
-    connect(clipboard, &QClipboard::dataChanged, [this]() {
+    connect(clipboard, &QClipboard::dataChanged, this, [this]() {
         // We need to store the clipboard ourselves to preserve external clipboard changes we would
         // not catch otherwise
         keepClipboard();